requirements.txt

aiogram==2.25.1
aiohttp==3.8.5
aiosqlite==0.19.0
APScheduler==3.9.1
python-dotenv==1.0.0
